// src/features/auth/infrastructure/api/sections-섹션/sections-섹션.ts
    export type sectionsControllerFindAllResponse200 = {
  data: SectionEntity[]
  status: 200
}

export type sectionsControllerFindAllResponseSuccess = (sectionsControllerFindAllResponse200) & {
  headers: Headers;
};
;

export type sectionsControllerFindAllResponse = (sectionsControllerFindAllResponseSuccess)