// src/features/auth/infrastructure/api/model/reminderEntity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Tickit Server API
 * Tickit 클라우드 서버 API 명세서
 * OpenAPI spec version: 1.0
 */

export interface ReminderEntity {
  /** 리마인더 고유 ID */
  id: number;
  /** 리마인더 내용 */
  text: string;
  /**
     * 알림 시간
     * @nullable
     */
  time: string | null;
  /** 하루 종일 여부 */
  isAllDay: boolean;
  /** 알림 발송 여부 */
  notified: boolean;
  /** 완료 여부 */
  done: boolean;
  /** 소속 섹션 ID */
  sectionId: string;
  /** 생성일 */
  createdAt: string;
  /** 수정일 */
  updatedAt: string;
}